Add `cargo-zng`
