+++ crates/cargo-zng/Cargo.toml
@@ -1,19 +1,27 @@
 [package]
-name = "zng-l10n-scraper"
-version = "0.3.1+deprecated"
+name = "cargo-zng"
+version = "0.1.0"
 authors = ["The Zng Project Developers"]
 edition = "2021"
 license = "Apache-2.0 OR MIT"
 readme = "README.md"
-description = "Command-line tool that scraps l10n text from Zng apps code."
-documentation = "https://zng-ui.github.io/doc/zng_l10n_scraper"
+description = "Zng project manager."
+documentation = "https://zng-ui.github.io/doc/zng_unique_id"
 repository = "https://github.com/zng-ui/zng"
 categories = ["gui"]
 keywords = ["gui", "ui", "user-interface", "zng"]
 
 [dependencies]
+# cli
+clap = { version = "*", features = ["derive"] }
+color-print = "0.3"
+
+# l10n
 rayon = "1.10"
 glob = "0.3"
 rustc_lexer = "0.1"
-clap = { version = "4.5", features = ["derive"] }
 fluent-pseudo = "0.3"
+
+
+[package.metadata.docs.rs]
+all-features = true
