+++ crates/cargo-zng/README.md
@@ -1,0 +1,24 @@
+<!--do doc --readme header-->
+This crate is part of the [`zng`](https://github.com/zng-ui/zng?tab=readme-ov-file#crates) project.
+
+Cargo extension for Zng project management. Create a new project from templates, collect localization strings, package
+the application for distribution.
+
+# Installation
+
+```console
+cargo install cargo-zng
+```
+
+# Usage
+
+```console
+# cargo zng --help
+
+todo!()
+
+```
+
+[documentation]: https://zng-ui.github.io/doc/zng/l10n/macro.l10n.html#scrap-template
+
+<!--do doc --readme features-->
